+++ nomos-da/network/core/src/protocols/sampling/mod.rs
@@ -48,7 +48,7 @@ mod test {
     #[tokio::test]
     #[expect(clippy::too_many_lines)]
     async fn test_sampling_two_peers() {
-        const MSG_COUNT: usize = 10;
+        const MSG_COUNT: u8 = 10;
         async fn test_sampling_swarm(
             mut swarm: Swarm<
                 SamplingBehaviour<
@@ -96,7 +96,7 @@ mod test {
                         debug!("{event:?}");
                     }
                 }
-                if res.len() == MSG_COUNT {
+                if res.len() == MSG_COUNT as usize {
                     break res;
                 }
             }
@@ -154,8 +154,8 @@ mod test {
         });
         tokio::time::sleep(Duration::from_secs(2)).await;
         for i in 0..MSG_COUNT {
-            request_sender_1.send((0, [i as u8; 32])).unwrap();
-            request_sender_2.send((0, [i as u8; 32])).unwrap();
+            request_sender_1.send((0, [i; 32])).unwrap();
+            request_sender_2.send((0, [i; 32])).unwrap();
         }
 
         let res1 = t1.await.unwrap();
